corrected Kruskal
