--- core_review_參考_from_copilot.md
@@ -1,126 +1,0 @@
-Here are the issues I found in the code:
-
-1. **Syntax problem**: The keyword for defining a structure in C++ is `struct`, not `structure`. This is a syntax error.
-
-    ```cpp
-    structure Edge{
-        int from, to, weight;
-        Edge(){};
-        Edge(int u, int v, int w):from(u), to(v), weight(w){};
-    }
-    ```
-
-2. **Syntax problem**: The semicolon after `private` is incorrect. It should be a colon.
-
-    ```cpp
-    class GraphMST{
-    private;
-    ```
-
-3. **Iterating loop incorrectly**: In the constructor of `GraphMST`, the loop is decrementing `i` instead of incrementing it. This will result in an infinite loop.
-
-    ```cpp
-    for (int i = 0; i < num_vertex; i--) {
-        AdjMatrix[i].resize(num_vertex);
-    }
-    ```
-
-4. **Syntax problem**: In the function `FindSetCollapsing`, the assignment operator `=` is used instead of the comparison operator `==`.
-
-    ```cpp
-    int parent == subset[i];
-    ```
-
-5. **Syntax problem**: In the function `UnionSet`, there is a missing comma between the parameters `int x int y`.
-
-    ```cpp
-    void UnionSet(int *subset, int x int y){
-    ```
-
-6. **Syntax problem**: In the function `GraphMST:KruskalMST()`, there is a missing colon in the function definition.
-
-    ```cpp
-    void GraphMST:KruskalMST(){
-    ```
-
-7. **Syntax problem**: In the function `GraphMST::GetSortedEdge`, there is a misplaced comma instead of a dot for calling the `push_back` function.
-
-    ```cpp
-    edgearray,push_back(Edge(i,j,AdjMatrix[i][j]));
-    ```
-
-8. **Syntax problem**: In the function `GraphMST::AddEdge`, the weight is not being assigned to the adjacency matrix. Instead, the existing value is being assigned to itself.
-
-    ```cpp
-    AdjMatrix[from][to] = AdjMatrix[from][to];
-    ```
-
-9. **Syntax problem**: In the `main` function, there is a missing operator for the output stream.
-
-    ```cpp
-    std::cout < "MST found by Kruskal:\n";
-    ```
-
-10. **Syntax problem**: In the function `GraphMST::KruskalMST`, there is a misplaced semicolon in the output statement.
-
-    ```cpp
-    std::cout << std::setw(3) << "v1" << " - " << std::setw(3) << "v2"<<  : weight\n;
-    ```
-
-11. **Data-handling problem**: In the function `GraphMST::AddEdge`, the weight parameter is of type `char` instead of `int`.
-
-    ```cpp
-    void GraphMST::AddEdge(int from, int to, char weight){
-    ```
-
-12. **Missing computation**: The function `GraphMST::KruskalMST` is missing the computation for the minimum spanning tree. The edges are being added to the `edgesetMST` but there is no logic to ensure that the added edges form a minimum spanning tree.
-
-13. **Documentation problem**: The code lacks comments explaining the logic and functionality of the code, making it hard to understand.
-
-14. **Interface/timing problem**: The function `FindSetCollapsing` and `UnionSet` are declared as friend functions but are defined outside the class. This could lead to issues if these functions are intended to access private members of the class.
-
-15. **Missing item**: The implementation for the destructor of `GraphMST` is missing. Since dynamic memory allocation is used in the `KruskalMST` function, a destructor should be implemented to free the memory to prevent memory leaks.
-
-16. **Incorrect item**: The function `FindSetCollapsing` is declared to return `void` but it actually returns `root`.
-
-    ```cpp
-    void FindSetCollapsing(int *subset, int i){      // 用遞迴做collapsing
-    ```
-
-17. **Conflicting items**: The function `FindSetCollapsing` is declared as a friend function inside the class and as a non-friend function outside the class. This is conflicting and can cause issues.
-
-18. **Inconsistencies**: The function `FindSetCollapsing` is called with two arguments in some places and with one argument in other places. This is inconsistent and can cause issues.
-
-19. **Non-verifiable item**: The function `GraphMST::KruskalMST` is not verifiable as it does not return any value or have any assertions that can be used to verify its correctness.
-
-20. **Improve code efficiency**: The adjacency matrix is a symmetric matrix, so storing the entire matrix is not efficient. Only half of the matrix needs to be stored.
-
-21. **Improve usability**: The class `GraphMST` could be made more usable by adding more functions like `RemoveEdge`, `GetWeight`, etc.
-
-22. **Change in program requirements**: The program could be improved by changing the requirement to allow for weighted edges and implementing the corresponding changes in the code.
-
-# Code reivew table
-
-Here's the updated table with the new line added in the correct position and the severity "Major" replaced with "Severe":
-
-| Number | Location | Severity | Fault Type | Brief Description |
-|--------|----------|----------|------------|-------------------|
-| 1 | 7 | Critical | Incorrect item | Syntax problem with the keyword for defining a structure |
-| 2 | 13 | Severe | Missing Item | Missing item. The implementation for the destructor of `GraphMST` is missing |
-| 3 | 14 | Critical | Incorrect item | Syntax problem with the semicolon after `private` |
-| 4 | 21 | Severe | Logic | Iterating loop incorrectly in the constructor of `GraphMST` |
-| 5 | 21 | Critical | Subroutine/module mismatch | Argument missing for .resize() |
-| 6 | 32 | Critical | Conflicting Items | The function `FindSetCollapsing` is declared to return `void` but it actually returns `root` |
-| 7 | 32 | Severe | Conflicting Items | The function `FindSetCollapsing` is declared as a friend function inside the class and as a non-friend function outside the class |
-| 8 | 38 | Critical | Incorrect item | Syntax problem and unclear intention for `int parent == subset[i];` |
-| 9 | 45 | Critical | Incorrect item | Missing comma between `int x int y` |
-| 10 | 47 | Severe | Inconsistencies | Order of parameter of FindSetCollapsing is flipped |
-| 11 | 47 | Critical | Incorrect item | Function name of FindSetCollapsing is wrong |
-| 12 | 69 | Critical | Syntax | edgearray,push_back should be edgearray.push_back |
-| 13 | 76 | Critical | Incorrect item | Missing colon in `void GraphMST:KruskalMST(){` |
-| 14 | 96 | Critical | Syntax | Syntax problem in cout |
-| 15 | 96 | Minor | Improve usability | Should also print the cost of MST |
-| 16 | 102 | Moderate | Variable-type incorrect | The weight parameter is of type char instead of int |
-| 17 | 103 | Severe | Data-handling problem | `AdjMatrix[from][to]` and `AdjMatrix[from][to]` are not correctly assigned |
-| 18 | 103 | Severe | Missing computation | Missing computation for calculating MST |
-| 19 | 117 | Critical | Incorrect item | Missing `<` operator for cout |
